backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
python-multipart==0.0.10
pillow==11.2.1
tensorflow>=2.16.0
numpy>=2.1.0 